can now set in the config file a launch command with --set-options
